Remove "_int" suffix from protocol submodules
